# Y-T-AUTO-master/Y-T-AUTO-master/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Y tế cho người bình thường API",
    "description": "API Backend cho ứng dụng hỗ trợ giải thích chỉ số xét nghiệm và gợi ý thực đơn, bài tập hỗ trợ.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/auth/phone": {
      "get": {
        "operationId": "getPhoneAccountStatus",
        "summary": "Get the authenticated account phone verification status",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Phone account status",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/PhoneAccountStatus"
                    },
                    "error": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "A valid application Bearer session is required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Phone status could not be read",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/oauth/exchange": {
      "post": {
        "operationId": "exchangeOAuthCallbackCode",
        "summary": "Exchange a single-use OAuth callback code for an application session",
        "security": [],
        "parameters": [
          {
            "name": "yte_oauth_binding",
            "in": "cookie",
            "required": true,
            "description": "HttpOnly browser correlation cookie issued when OAuth starts.",
            "schema": {
              "type": "string",
              "minLength": 43,
              "maxLength": 43,
              "pattern": "^[A-Za-z0-9_-]{43}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "code",
                  "intent"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "minLength": 43,
                    "maxLength": 43,
                    "pattern": "^[A-Za-z0-9_-]{43}$"
                  },
                  "intent": {
                    "type": "string",
                    "enum": [
                      "LOGIN",
                      "REGISTER",
                      "LINK"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OAuth session created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "success",
                    "data",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/OAuthSession"
                    },
                    "error": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The callback code or authorization state is invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The OAuth identity requires explicit linking or conflicts with another account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The application session could not be created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "The OAuth provider failed to complete authentication",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "OAuthSession": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "userId",
          "token",
          "intent"
        ],
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 8192,
            "pattern": "^\\S+$"
          },
          "intent": {
            "type": "string",
            "enum": [
              "LOGIN",
              "REGISTER",
              "LINK"
            ]
          }
        }
      },
...
      "PhoneAccountStatus": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "phoneVerified",
          "maskedPhone"
        ],
        "properties": {
          "phoneVerified": {
            "type": "boolean"
          },
          "maskedPhone": {
            "type": [
              "string",
              "null"
            ],
            "pattern": "^\\+[1-9][0-9]{0,2}\\*{4,12}[0-9]{2,4}$"
          }
        },
        "oneOf": [
          {
            "properties": {
              "phoneVerified": {
                "const": false
              },
              "maskedPhone": {
                "type": "null"
              }
            }
          },
          {
            "properties": {
              "phoneVerified": {
                "const": true
              },
              "maskedPhone": {
                "type": "string",
                "pattern": "^\\+[1-9][0-9]{0,2}\\*{4,12}[0-9]{2,4}$"
              }
            }
          }
        ]
      },
...
      "ApiErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "success",
          "data",
          "error"
        ],
        "properties": {
          "success": {
            "const": false
          },
          "data": {
            "type": "null"
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128,
                "pattern": "^[A-Z][A-Z0-9_]*$"
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 1000,
                "pattern": "\\S"
              },
              "details": {
                "type": "array",
                "minItems": 1,
                "items": {
                  "$ref": "#/components/schemas/ValidationIssue"
                }
              }
            }
          }
        }
      },
      "ValidationIssue": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "code",
          "message",
          "path"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          },
          "path": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          }
        }
      },
...
    }
  }
}